Update show.sql
Made it so that show_id is automatically incrementing itself whenever data is inserting inside the show_table. Also we created stored procedures where you can call either anime_Insert or cartoon_Insert and provide it with 3 values (that is related to the type of show they are) in which it will properly insert the respective type into the table. Designed a trigger that will check on update in which will see if the new name is already inside the show_table, and if it is then print out a message specifying that name is already there and what type it is. Designed another trigger for insert that checks if the name already exists and if it does print out that the name is already taken.